+++ spec/decanter/parser/float_parser_spec.rb
@@ -3,23 +3,23 @@
 require 'spec_helper'
 
 describe Decanter::Parser::FloatParser do
-  subject { described_class.parse(name, arg) }
+  subject { described_class.parse(arg) }
   let(:name) { :foo }
 
   describe '#parse' do
     context 'with an integer-like string ' do
       let(:arg) { '1' }
-      it { is_expected.to match(name => 1.0) }
+      it { is_expected.to eq 1.0 }
     end
 
     context 'with an float-like string ' do
       let(:arg) { '1.1' }
-      it { is_expected.to match(name => 1.1) }
+      it { is_expected.to eq 1.1 }
     end
 
     context 'with nil' do
       let(:arg) { nil }
-      it { is_expected.to match(name => nil) }
+      it { is_expected.to be_nil }
     end
 
     context 'with a duff string' do