+++ spec/decanter/parser/parser_core_spec.rb
@@ -2,13 +2,13 @@
 
 require 'spec_helper'
 
-describe 'Core' do
+describe Decanter::Parser::Core do
   let(:core) { Class.new { include Decanter::Parser::Core } }
 
   describe '#parser' do
     it 'sets class variable @parser to equal the block' do
       parser = proc {}
-      core.parser &parser
+      core.parser(&parser)
       expect(core.instance_variable_get(:@parser)).to eq parser
     end
   end
@@ -23,13 +23,13 @@ describe 'Core' do
   describe '#parse' do
     context 'for blank value' do
       it 'returns the value' do
-        expect(core.parse('first_name', nil)).to match('first_name' => nil)
+        expect(core.parse(nil)).to be_nil
       end
     end
 
     context 'for empty string' do
       it 'returns nil' do
-        expect(core.parse('first_name', '')).to match('first_name' => nil)
+        expect(core.parse('')).to be_nil
       end
     end
 
@@ -43,12 +43,12 @@ describe 'Core' do
 
       context 'when type is allowed' do
         it 'returns the value' do
-          expect(core.parse('first_name', 'foo')).to match('first_name' => 'foo')
+          expect(core.parse('foo')).to eq 'foo'
         end
 
         it 'does not call the parser' do
           allow(core).to receive(:_parse)
-          core.parser &parser
+          core.parser(&parser)
           core.parse('first_name', ['foo'])
           expect(core).to_not have_received(:_parse)
         end
@@ -58,11 +58,11 @@ describe 'Core' do
         context 'when a parser is defined' do
           it 'calls the parser' do
             allow(core).to receive(:_parse)
-            core.parser &parser
-            core.parse('first_name', 5, required: true)
+            core.parser(&parser)
+            core.parse(5, required: true)
             expect(core)
               .to have_received(:_parse)
-              .with('first_name', 5, required: true)
+              .with(5, required: true)
           end
         end
       end