+++ spec/decanter/parser/string_parser_spec.rb
@@ -2,10 +2,11 @@
 
 require 'spec_helper'
 
-describe 'StringParser' do
+describe Decanter::Parser::StringParser do
+  subject { described_class.parse(arg) }
+  let(:arg) { 8 }
+
   describe '#parse' do
-    it 'returns a string' do
-      expect(Decanter::Parser::StringParser.parse(:foo, 8)).to match(foo: '8')
-    end
+    it { is_expected.to eq '8' }
   end
 end