+++ CHANGELOG.md
@@ -4,6 +4,8 @@
   * Removed `_new`, `_update` etc calls from models, replacing them with a
     `decant` method which can be called from with a controller, if
     `Decanter::Decant` is included. (!1)
+  * Refactored Parser#parse not to require the name option (!9)
+  * Refactored Decanter::Core#parse (!9)
   * Added TimeParser, clarified DateParser, DateTimeParser (!5)
   * Altered `FloatParser` and `IntegerParser` to use `Float()` and `Integer()`
     respectively (#1,!7)